fix(structuredlist): Change to composed, add carbon version
